@@ -4,25 +4,26 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.1.0+9
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
   # visuals 
-  cupertino_icons: ^1.0.0
+  cupertino_icons: ^1.0.3
   # state management
-  flutter_bloc: ^6.1.1
-  hydrated_bloc: ^6.1.0
-  equatable: ^1.2.5
-  shared_preferences: ">=0.5.12+4 <2.0.0" # save user preferences
+  flutter_bloc: ^7.2.0
+  hydrated_bloc: ^7.0.1
+  equatable: ^2.0.3
+  shared_preferences: ^2.0.7
+  path_provider: ^2.0.3
   # service to save instances
-  get_it: ^5.0.0
+  get_it: ^7.2.0
   # dependency to play sounds
-  assets_audio_player: ^2.0.13+8
+  assets_audio_player: ^3.0.3+6
 
 dev_dependencies:
-  flutter_launcher_icons: ^0.8.1
+  flutter_launcher_icons: ^0.9.2
   flutter_test:
     sdk: flutter
   
