@@ -1,4 +1,3 @@
-import 'package:chessclock/misc/utils/color_utils.dart';
 import 'package:flutter/material.dart';
 
 /// Message Widget to show [SnackBar]s in the app using a [globalKey].
@@ -6,26 +5,23 @@ class MessageWidget extends StatefulWidget {
   static final GlobalKey<MessageWidgetState> globalKey =
       GlobalKey<MessageWidgetState>();
 
-  final Widget child;
-  MessageWidget({@required this.child}) : super(key: globalKey);
+  final Widget? child;
+
+  MessageWidget({required this.child}) : super(key: globalKey);
 
   @override
   MessageWidgetState createState() => MessageWidgetState();
 }
 
 class MessageWidgetState extends State<MessageWidget> {
-  final GlobalKey<ScaffoldState> _scaffoldKey = GlobalKey<ScaffoldState>();
+  final GlobalKey<ScaffoldMessengerState> _messengerKey =
+      GlobalKey<ScaffoldMessengerState>();
 
-  void show(String message, [SnackBarAction action]) {
-    final _theme = Theme.of(context);
-    final _contrast = getContrastColor(_theme.accentColor);
-    final _textStyle = _theme.textTheme.bodyText2.copyWith(color: _contrast);
-
-    _scaffoldKey.currentState.hideCurrentSnackBar();
-    _scaffoldKey.currentState.showSnackBar(
+  void show(String message, [SnackBarAction? action]) {
+    _messengerKey.currentState!.hideCurrentSnackBar();
+    _messengerKey.currentState!.showSnackBar(
       SnackBar(
-        content: Text(message, style: _textStyle),
-        backgroundColor: _theme.accentColor,
+        content: Text(message),
         action: action,
       ),
     );
@@ -33,6 +29,9 @@ class MessageWidgetState extends State<MessageWidget> {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(key: _scaffoldKey, body: widget.child);
+    return ScaffoldMessenger(
+      key: _messengerKey,
+      child: Scaffold(body: widget.child),
+    );
   }
 }
