@@ -1,7 +1,7 @@
 class AppThemeModel {
-  String name;
-  int primaryColour;
-  int accentColor;
+  String? name;
+  int? primaryColour;
+  int? accentColor;
 
   AppThemeModel({
     this.name,
