@@ -4,23 +4,23 @@ abstract class AddClockEvent extends Equatable {
   const AddClockEvent();
 
   @override
-  List<Object> get props => [];
+  List<Object?> get props => [];
 }
 
 class AddCustomClock extends AddClockEvent {
-  final ChessClock chessClock;
+  final ChessClock? chessClock;
 
   const AddCustomClock({this.chessClock});
 
   @override
-  List<Object> get props => [chessClock];
+  List<Object?> get props => [chessClock];
 }
 
 class RemoveCustomClock extends AddClockEvent {
-  final int index;
+  final int? index;
 
   const RemoveCustomClock({this.index});
 
   @override
-  List<Object> get props => [index];
+  List<Object?> get props => [index];
 }
