@@ -4,13 +4,13 @@ Color getContrastColor(Color color) {
   return color.computeLuminance() > .5 ? Colors.black : Colors.white;
 }
 
-Color lighten(Color color, [double amount = .1]) {
+Color? lighten(Color? color, [double amount = .1]) {
   assert(amount >= 0 && amount <= 1);
 
   // checks if color is as dark as it gets
   if (color == Color(0xff000000)) return color;
 
-  final hsl = HSLColor.fromColor(color);
+  final hsl = HSLColor.fromColor(color!);
   final hslLight = hsl.withLightness((hsl.lightness + amount).clamp(0.0, 1.0));
 
   return hslLight.toColor();
